Compile and run! Huzzah, but still broked
There appears to be a problem with swapCursor... I'll look into it,
hopefully won't be hard to fix.... still does not display any of the
results and what not. I fear that the adapter or whatever does not
work... I have a new sense of purpose with the project now thanks to
this breakthrough!